requirements.txt
playwright==1.58.0
pyotp==2.9.0
python-dotenv==1.2.2
beautifulsoup4==4.8.2
requests==2.32.3
pytest==9.0.2
lxml==5.4.0
